feat: retune Trident world generation for 64x64
Defaults picked by sweeping ~2500 worlds (docs/TRIDENT_MAP_SELECTION.md):
frequency 5.2, octaves 5, persistence 0.6, water 0.60, equator 0.40,
mountains 0.08, hills 0.57, polar block from 82, settlementsKept 5.
Adds noise aspect correction, rational underground scales (3/4, 3/8)
and shaft spacing (one end per hex, 2 apart, never in lakes).

Trident only: dev has a different generator.

+++ unittest/world.cpp
@@ -24,10 +24,10 @@ void Game::CreateWorld() {
 int ARegionList::GetRegType( ARegion *pReg ) { return 0; }
 
 // Unit test levels are unscaled.
-int ARegionList::GetLevelXScale(int level) { return 1; }
+double ARegionList::GetLevelXScale(int level) { return 1.0; }
 
 // Unit test levels are unscaled.
-int ARegionList::GetLevelYScale(int level) { return 1; }
+double ARegionList::GetLevelYScale(int level) { return 1.0; }
 
 // Unit test regions are fully connected
 int ARegionList::CheckRegionExit(ARegion *pFrom, ARegion *pTo ) { return 1; }
